Rearrange slicer behaviour and slicers.
